requirements.txt
pandas==1.4.4
numpy==1.23.5
matplotlib==3.5.2
keras==2.10.0; sys_platform == "darwin"
keras==2.11.0; sys_platform == "windows" or sys_platform == "linux"
sklearn-pandas==1.8.0
absl-py==1.3.0
astor==0.8.1
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
flatbuffers==22.12.6
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
joblib==1.2.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
preproc==0.0.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
tensorboard==2.10.1; sys_platform == "darwin"
tensorboard==2.11; sys_platform == "windows" or sys_platform == "linux"
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0; sys_platform == "darwin"
tensorflow-estimator==2.11.0; sys_platform == "windows" or sys_platform == "linux"
tensorflow-macos==2.10.0; sys_platform == "darwin"
tensorflow==2.11.0; sys_platform == "windows" or sys_platform == "linux"
termcolor==2.1.1
threadpoolctl==3.1.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
